moderation_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy>=1.26.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
transformers>=4.35.0
torch>=2.1.0
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.0 